Add FMSManager and delegate simulation logic
